Add files via upload
In this project, i implemented a new scheduler into xv6. It is called a lottery
scheduler The basic idea is simple: assign each running
process a slice of the processor based in proportion to the number of tickets it has;
the more tickets a process has, the more it runs. Each time slice, a randomized
lottery determines the winner of the lottery; that winning process is the one that
runs for that time slice.